sagemaker/detr/requirements.txt
torch
transformers
timm
Pillow
boto3